+++ CrossPlatform/milkbar_qt_gui.py
@@ -432,7 +432,7 @@ class MilkBarWindow(QMainWindow):
         self.graphics_process.setArguments(["--mlc", str(backend.data_directory() / "cemu" / "mlc01"),
                                             "--graphic-packs", title_id])
         environment = QProcessEnvironment()
-        for key, value in backend.external_process_environment().items():
+        for key, value in backend.cemu_process_environment(cemu).items():
             environment.insert(key, value)
         environment.insert("MILKBAR_CEMU_DATA_DIR", str(config_dir))
         self.graphics_process.setProcessEnvironment(environment)